requirements.txt
# Core Dependencies
streamlit==1.31.0
langchain==0.1.10
langchain-community==0.0.25
langchain-openai==0.0.7

# OpenAI
openai==1.12.0

# Vector Databases
chromadb==0.4.22
faiss-cpu==1.7.4

# YouTube Processing
youtube-transcript-api==0.6.2
pytube==15.0.0

# NLP and Text Processing
tiktoken==0.6.0
nltk==3.8.1
spacy==3.7.4

# Data Processing
numpy==1.26.4
pandas==2.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0

# Additional LangChain Components
langchain-core==0.1.28
langsmith==0.1.1

# Embeddings
sentence-transformers==2.3.1

# Environment
python-multipart==0.0.9
